+++ CHANGES.md
@@ -5,6 +5,9 @@
 - Fix: Add an `__iter__` method to python iterators.
   (Fixed by Laurent Mazare, https://github.com/thierry-martinez/pyml/pull/47)
 
+- Add `Py.Seq.{of_seq_map, to_seq_map, unsafe_to_seq_map, of_list,
+  of_list_map}` functions.
+
 # 2020-02-22
 
 - Fix: do not fail if GIL functions are unavailable